refactor : 거래내역에서 사용한 메서드 util 함수로 따로 관리
